Add workaround for proxy basic auth not working with https request
With the Mac's Network settings configured to use a basic auth proxy
for http and https requests, the username and password also stored
with the network settings, the https request send by the WebView
would fail with a proxy auth failure. Error kCFErrorDomainCFNetwork
310 which maps to kCFErrorHTTPSProxyConnectionFailure. The sign in
dialog would just show a progress spinner icon.

As a workaround, if the machine uses a proxy, that supports basic
auth, and has credentials stored in the key chain, the
Proxy-Authorization header will be added to the request with the
basic auth credentials.